backend/requirements.txt
fastapi
uvicorn[standard]
strawberry-graphql
pymysql
redis
pymongo